requirements.txt
Flask==3.0.0
python-docx==1.1.0
Pillow==10.1.0
requests==2.31.0
urllib3==2.0.7
Werkzeug==3.0.1
pyparsing==3.1.1
docutils==0.20.1
lxml==4.9.3
Jinja2==3.1.2
pandas==2.1.2